requirements.txt
flask
gevent
requests
pillow
